backend/ml.py
from PIL import Image
import os
from config import Config

def run_biopsy_yolo(image_path):
    """
    Run biopsy image through YOLO model (or fake mode).
    Returns dict with cancer prediction and confidence.
    """
    if Config.USE_FAKE_ML:
        return {"has_cancer": True, "confidence": 0.91}
    
    # Real model integration placeholder
    img = Image.open(image_path)
    # TODO: Pass img to your YOLO model and get prediction
    return {"has_cancer": True, "confidence": 0.9}


def analyze_mammograms(image_paths):
    """
    Analyze a list of mammogram image paths.
    Returns prediction, risk score, and tumor size trend.
    """
    if Config.USE_FAKE_ML:
        return {
            "prediction": "High risk",
            "score": 0.78,
            "trend": [
                {"t": 1, "size": 3.2},
                {"t": 2, "size": 3.6},
                {"t": 3, "size": 3.9}
            ]
        }

    # Real model analysis placeholder
    # TODO: Integrate your YOLO/mammogram analysis pipeline
    return {"prediction": "Pending", "score": 0.0, "trend": []}
