backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.32
alembic==1.13.2
Werkzeug==3.0.3
Pillow==10.4.0
flask
flask-cors
flask-jwt-extended
flask-pymongo
dnspython
